scan_wrapper renamed to run_tasks
