+++ obi_one/__init__.py
@@ -5,7 +5,7 @@ from obi_one.core.exception import OBIONEError
 from obi_one.core.info import Info
 from obi_one.core.path import NamedPath
 from obi_one.core.scan_config import ScanConfig
-from obi_one.core.scan_wrapper import (
+from obi_one.core.run_tasks import (
     run_task_for_single_config,
     run_task_for_single_configs,
     run_tasks_for_generated_scan,
